+++ mis_builder/models/mis_report_instance.py
@@ -5,6 +5,8 @@
 import datetime
 import logging
 
+from dateutil.relativedelta import relativedelta
+
 from odoo import api, fields, models, _
 from odoo.exceptions import ValidationError, UserError
 
@@ -111,6 +113,26 @@ class MisReportInstancePeriod(models.Model):
                 record.date_from = fields.Date.to_string(date_from)
                 record.date_to = fields.Date.to_string(date_to)
                 record.valid = True
+            elif record.mode == MODE_REL and record.type == 'm':
+                date_from = d.replace(day=1)
+                date_from = date_from + \
+                    relativedelta(months=record.offset)
+                date_to = date_from + \
+                    relativedelta(months=record.duration-1) + \
+                    relativedelta(day=31)
+                record.date_from = fields.Date.to_string(date_from)
+                record.date_to = fields.Date.to_string(date_to)
+                record.valid = True
+            elif record.mode == MODE_REL and record.type == 'y':
+                date_from = d.replace(month=1, day=1)
+                date_from = date_from + \
+                    relativedelta(years=record.offset)
+                date_to = date_from + \
+                    relativedelta(years=record.duration-1)
+                date_to = date_to.replace(month=12, day=31)
+                record.date_from = fields.Date.to_string(date_from)
+                record.date_to = fields.Date.to_string(date_to)
+                record.valid = True
             elif record.mode == MODE_REL and record.type == 'date_range':
                 date_range_obj = record.env['date.range']
                 current_periods = date_range_obj.search(
@@ -119,8 +141,12 @@ class MisReportInstancePeriod(models.Model):
                      ('date_end', '>=', d),
                      '|',
                      ('company_id', '=', False),
-                     ('company_id', '=', report.company_id.id)])
+                     ('company_id', 'in',
+                      record.report_instance_id.query_company_ids.ids)])
                 if current_periods:
+                    # TODO we take the first date range we found as current
+                    #      this may be surprising if several companies
+                    #      have overlapping date ranges with different dates
                     current_period = current_periods[0]
                     all_periods = date_range_obj.search(
                         [('type_id', '=', current_period.type_id.id),
@@ -148,6 +174,8 @@ class MisReportInstancePeriod(models.Model):
     type = fields.Selection(
         [('d', _('Day')),
          ('w', _('Week')),
+         ('m', _('Month')),
+         ('y', _('Year')),
          ('date_range', _('Date Range'))],
         string='Period type'
     )
@@ -208,9 +236,10 @@ class MisReportInstancePeriod(models.Model):
         domain=[('field_id.name', '=', 'debit'),
                 ('field_id.name', '=', 'credit'),
                 ('field_id.name', '=', 'account_id'),
-                ('field_id.name', '=', 'date')],
+                ('field_id.name', '=', 'date'),
+                ('field_id.name', '=', 'company_id')],
         help="A 'move line like' model, ie having at least debit, credit, "
-             "date and account_id fields.",
+             "date, account_id and company_id fields.",
     )
     source_sumcol_ids = fields.One2many(
         comodel_name='mis.report.instance.period.sum',
@@ -334,9 +363,10 @@ class MisReportInstance(models.Model):
                 record.pivot_date = fields.Date.context_today(record)
 
     @api.model
-    def _default_company(self):
-        return self.env['res.company'].\
-            _company_default_get('mis.report.instance')
+    def _default_company_id(self):
+        default_company_id = self.env['res.company'].\
+            _company_default_get('mis.report.instance').id
+        return default_company_id
 
     _name = 'mis.report.instance'
 
@@ -362,10 +392,35 @@ class MisReportInstance(models.Model):
                                    string='Target Moves',
                                    required=True,
                                    default='posted')
-    company_id = fields.Many2one(comodel_name='res.company',
-                                 string='Company',
-                                 default=_default_company,
-                                 required=True)
+    company_id = fields.Many2one(
+        comodel_name='res.company',
+        string='Company',
+        default=_default_company_id,
+        required=True,
+    )
+    multi_company = fields.Boolean(
+        string='Multiple companies',
+        help="Check if you wish to specify "
+             "children companies to be searched for data.",
+        default=False,
+    )
+    company_ids = fields.Many2many(
+        comodel_name='res.company',
+        string='Companies',
+        help="Select companies for which data will be searched.",
+    )
+    query_company_ids = fields.Many2many(
+        comodel_name='res.company',
+        compute='_compute_query_company_ids',
+        help="Companies for which data will be searched.",
+    )
+    currency_id = fields.Many2one(
+        comodel_name='res.currency',
+        string='Currency',
+        help="Select target currency for the report. "
+             "Required if companies have different currencies.",
+        required=False,
+    )
     landscape_pdf = fields.Boolean(string='Landscape PDF')
     comparison_mode = fields.Boolean(
         compute="_compute_comparison_mode",
@@ -376,6 +431,24 @@ class MisReportInstance(models.Model):
     date_from = fields.Date(string="From")
     date_to = fields.Date(string="To")
     temporary = fields.Boolean(default=False)
+
+    @api.onchange('company_id', 'multi_company')
+    def _onchange_company(self):
+        if self.company_id and self.multi_company:
+            self.company_ids = self.env['res.company'].search([
+                ('id', 'child_of', self.company_id.id),
+            ])
+        else:
+            self.company_ids = False
+
+    @api.multi
+    @api.depends('multi_company', 'company_id', 'company_ids')
+    def _compute_query_company_ids(self):
+        for rec in self:
+            if rec.multi_company:
+                rec.query_company_ids = rec.company_ids or rec.company_id
+            else:
+                rec.query_company_ids = rec.company_id
 
     @api.multi
     def save_report(self):
@@ -584,7 +657,8 @@ class MisReportInstance(models.Model):
         is guaranteed to be the id of the mis.report.instance.period.
         """
         self.ensure_one()
-        aep = self.report_id._prepare_aep(self.company_id)
+        aep = self.report_id._prepare_aep(
+            self.query_company_ids, self.currency_id)
         kpi_matrix = self.report_id.prepare_kpi_matrix()
         for period in self.period_ids:
             description = None
@@ -615,7 +689,7 @@ class MisReportInstance(models.Model):
         account_id = arg.get('account_id')
         if period_id and expr and AEP.has_account_var(expr):
             period = self.env['mis.report.instance.period'].browse(period_id)
-            aep = AEP(self.company_id)
+            aep = AEP(self.query_company_ids, self.currency_id)
             aep.parse_expr(expr)
             aep.done_parsing()
             domain = aep.get_aml_domain_for_expr(