+++ mis_builder_budget/__manifest__.py
@@ -6,7 +6,7 @@
     'name': 'MIS Builder Budget',
     'summary': """
         Create budgets for MIS reports""",
-    'version': '10.0.1.1.1',
+    'version': '10.0.3.1.0',
     'license': 'AGPL-3',
     'author': 'ACSONE SA/NV,'
               'Odoo Community Association (OCA)',